Moved checkBookingOverlap(...) from SearchResource to DriveResource. checkBookingOverlap(...) is now a private method, and therefore cannot be tested directly. Moved getDrivesAssociatedWithUser(int) = helper method to checkBookingOverlap(...).
Fixes #196
