+++ src/main/java/se/lth/base/server/rest/DriveResource.java
@@ -9,8 +9,8 @@ import javax.ws.rs.container.ContainerRequestContext;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response.Status;
-import java.net.URISyntaxException;
 import java.util.ArrayList;
+import java.util.Iterator;
 import java.util.List;
 
 @Path("drive")
@@ -108,10 +108,15 @@ public class DriveResource {
     @Consumes(MediaType.APPLICATION_JSON + ";charset=utf-8")
     public DriveUser addUserToDrive(@PathParam("driveId") int driveId, DriveUser driveUser) {
         if (driveDao.getDrive(driveId).getCarNumberOfSeats() > driveUserDao.getNumberOfUsersInDrive(driveId)) {
-            return driveUserDao.addDriveUser(driveId, user.getId(), driveUser.getStart(), driveUser.getStop(), !IS_DRIVER, !IS_ACCEPTED, !IS_RATED);
+            List<Drive> collidingDrives = checkBookingOverlap(driveId, driveUser.getStart(), driveUser.getStop());
+            if (collidingDrives.isEmpty()) {
+                return driveUserDao.addDriveUser(driveId, user.getId(), driveUser.getStart(), driveUser.getStop(), !IS_DRIVER, !IS_ACCEPTED, !IS_RATED);
+            } else {
+                throw new WebApplicationException("This drive overlaps in time with other booked drives", Status.PRECONDITION_FAILED);
+            }
+        } else {
+            throw new WebApplicationException("No available seats left", Status.PRECONDITION_FAILED);
         }
-
-        throw new WebApplicationException("No available seats left", Status.PRECONDITION_FAILED);
     }
 
     @Path("{driveId}/user/{userId}")
@@ -209,5 +214,88 @@ public class DriveResource {
             List<DriveReport> reports = driveReportDao.getDriveReportsForDrive(d.getDriveId());
             driveWraps.add(new DriveWrap(d, milestones, users, reports));
         }
+    }
+
+    /**
+     * Checks if it is possible for a user to book a specific Drive.
+     *
+     * @param driveId the id of the drive that one wishes to see if it is possible to book
+     * @param start   at what milestone the user will be starting at
+     * @param stop    at what milestone the user will stop at
+     * @return a list of drives associated with this user, conflicting with the entered drive
+     */
+    private List<Drive> checkBookingOverlap(@PathParam("driveId") int driveId, @PathParam("start") String start, @PathParam("stop") String stop) {
+        Drive drive = driveDao.getDrive(driveId);
+        List<DriveMilestone> driveMilestones = driveMilestoneDao.getMilestonesForDrive(driveId);
+        // Add drive start and stop as milestones to list
+        driveMilestones.add(0, new DriveMilestone(-1, -1, drive.getStart(), drive.getDepartureTime()));
+        driveMilestones.add(new DriveMilestone(-1, -1, drive.getStop(), drive.getArrivalTime()));
+
+        // Find the time that the user would be busy if user participated in drive
+        long busyTimeStart = -1;
+        long busyTimeEnd = -1;
+
+        for (DriveMilestone dm : driveMilestones) {
+            if (dm.getMilestone().toLowerCase().trim().equals(start.toLowerCase().trim())) {
+                busyTimeStart = dm.getDepartureTime();
+            } else if (dm.getMilestone().toLowerCase().trim().equals(stop.toLowerCase().trim())) {
+                busyTimeEnd = dm.getDepartureTime();
+            }
+        }
+
+        // Check for time conflicts with other already booked drives/trips
+        List<Drive> associatedDrives = getDriveAssociatedWithUser(user.getId());
+        Iterator<Drive> driveIterator = associatedDrives.iterator();
+        while (driveIterator.hasNext()) {
+            Drive d = driveIterator.next();
+            List<DriveMilestone> tempDriveMilestones = driveMilestoneDao.getMilestonesForDrive(d.getDriveId());
+            // Add drive start and stop as milestones to list
+            tempDriveMilestones.add(0, new DriveMilestone(-1, -1, d.getStart(), d.getDepartureTime()));
+            tempDriveMilestones.add(new DriveMilestone(-1, -1, d.getStop(), d.getArrivalTime()));
+
+            // Find the time that the user is busy in this drive
+            long tempBusyTimeStart = -1;
+            long tempBusyTimeEnd = -1;
+
+            DriveUser driveUser = driveUserDao.getDriveUser(d.getDriveId(), user.getId());
+
+            for (DriveMilestone dm : tempDriveMilestones) {
+                if (dm.getMilestone().toLowerCase().trim().equals(driveUser.getStart().toLowerCase().trim())) {
+                    tempBusyTimeStart = dm.getDepartureTime();
+                } else if (dm.getMilestone().toLowerCase().trim().equals(driveUser.getStop().toLowerCase().trim())) {
+                    tempBusyTimeEnd = dm.getDepartureTime();
+                }
+            }
+
+            if (tempBusyTimeEnd < busyTimeStart || tempBusyTimeStart > busyTimeEnd) {
+                // No conflict
+                driveIterator.remove();
+            }
+        }
+
+        // Return all booked drives that are in conflict with the new booking
+        return associatedDrives;
+    }
+
+    private List<Drive> getDriveAssociatedWithUser(int userId) {
+        List<Drive> drives = driveDao.getDrives();
+        Iterator<Drive> iterator = drives.iterator();
+        while (iterator.hasNext()) {
+            Drive drive = iterator.next();
+            List<DriveUser> driveUsers = driveUserDao.getDriveUsersForDrive(drive.getDriveId());
+            for (int i = 0; i < driveUsers.size(); i++) {
+                DriveUser driveUser = driveUsers.get(i);
+                if (driveUser.getUserId() == userId) {
+                    // User is associated with this drive, keep drive in list
+                    break;
+                }
+                if (i + 1 == driveUsers.size()) {
+                    // User is not associated with this drive, remove drive from list
+                    iterator.remove();
+                }
+            }
+
+        }
+        return drives;
     }
 }
