+++ src/test/java/se/lth/base/server/rest/SearchResourceTest.java
@@ -284,40 +284,4 @@ public class SearchResourceTest extends BaseResourceTest {
 
     }
     */
-
-    /**
-     * Make long booking 2018-10-20 12.00 - 18.00
-     * Try to att user to drive 1 (2018-10-20 12.00-12.30)
-     * Should be conflict
-     */
-    @Test
-    public void bookingOverlap() {
-        // Data access objects
-        DriveDataAccess driveDao = new DriveDataAccess(Config.instance().getDatabaseDriver());
-        DriveMilestoneDataAccess driveMilestoneDao = new DriveMilestoneDataAccess(Config.instance().getDatabaseDriver());
-        DriveUserDataAccess driveUserDao = new DriveUserDataAccess(Config.instance().getDatabaseDriver());
-
-        // Create long lasting drive
-        // Drive 4
-        long timestamp4 = new Timestamp(2018 - 1900, 10, 20, 12, 0, 0, 0).getTime();
-        long timestamp4_2 = new Timestamp(2018 - 1900, 10, 20, 18, 0, 0, 0).getTime();
-        Drive drive4 = driveDao.addDrive(new Drive(-1, "A", "F", timestamp4, timestamp4_2, "Comment", "x", "x", "x", "x", 1, 1, false, false, false));
-        int drive4Id = drive4.getDriveId();
-        long timestamp4_1 = new Timestamp(2018 - 1900, 10, 20, 12, 30, 0, 0).getTime();
-        driveMilestoneDao.addMilestone(drive4Id, "B", timestamp4_1);
-
-        // Add user to long lasting drive
-        driveUserDao.addDriveUser(drive4Id, user3Id, "A", "B", false, false, false);
-
-        login(SEARCH_TEST_CREDENTIALS_1);
-
-        // Check for overlaps
-        // We actually receive a List<LinkedTreeMap<String, Object>>
-        List<Drive> response = target("search")
-                .path("checkbookingoverlap/" + drive1Id + "/A/C")
-                .request()
-                .get(List.class);
-
-        Assert.assertEquals(1, response.size());
-    }
 }
